Orchestrator-eu/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
aiohttp>=3.9.0
redis>=5.0.0
pyyaml>=6.0.1
pydantic>=2.5.0
numpy>=1.24.0
twilio>=8.0.0
python-multipart>=0.0.6
tenacity>=8.2.0
kubernetes>=28.1.0
httpx>=0.25.0









